fix(apps): app 모듈 파일 이동
